+++ Sources/StreamChat/APIClient/Endpoints/ChannelEndpoints.swift
@@ -177,12 +177,17 @@ extension Endpoint {
         cid: ChannelId,
         members: [MemberInfoRequest],
         hideHistory: Bool,
+        hideHistoryBefore: Date? = nil,
         messagePayload: MessageRequestBody? = nil
     ) -> Endpoint<EmptyResponse> {
         var body: [String: AnyEncodable] = [
-            "add_members": AnyEncodable(members),
-            "hide_history": AnyEncodable(hideHistory)
+            "add_members": AnyEncodable(members)
         ]
+        if let hideHistoryBefore = hideHistoryBefore {
+            body["hide_history_before"] = AnyEncodable(hideHistoryBefore)
+        } else {
+            body["hide_history"] = AnyEncodable(hideHistory)
+        }
         if let messagePayload = messagePayload {
             body["message"] = AnyEncodable(messagePayload)
         }